requirements.txt
fredapi==0.5.0
lightgbm==3.3.5
matplotlib==3.7.0
numpy==1.24.2
pandas==1.5.3
pandas_datareader==0.10.0
python-dotenv==0.21.1
scikit_learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
setuptools==67.4.0
shap==0.41.0
statsmodels==0.13.5
yahoo_fin==0.8.9.1
yfinance==0.2.12
